requirements.txt
asgiref==3.11.0
boto3==1.34.44
botocore==1.34.162
dj-database-url==2.1.0
Django==5.1.2
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
django-unfold==0.72.0
djangorestframework==3.14.0
gunicorn==21.2.0
jmespath==1.0.1
packaging==25.0
Pillow==12.1.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.5.1
